feat(storage): implement in-memory storage
In-memory works as a cache backend for a Toml storage and for easier testing of the storage implementations. A Toml file is essentially just parsed an kept in-memory, every change is done in that storage then and at the end of the operations it is being synced back to disk.

This is useful, because we can test our storage implementations more easily by just working with the in-memory storage and we can easily spit out a backup file at every stage, if we want to.

Signed-off-by: simonsan <14062932+simonsan@users.noreply.github.com>