chore: add release-plz configuration
We need to configure the way we do releases for the pace-rs as the way release-plz currently does it, doesn't play well with cargo-dist. When release-plz pushed git tags and creates release from these tags, the cargo-dist workflow isn't being run and doesn't build any binaries. So instead of deleting the tag and release that is being created by release-plz and pushing our own tag again, we just don't create any tag and release for pace-rs from release-plz's side and just push a tag on our own.

Signed-off-by: simonsan <14062932+simonsan@users.noreply.github.com>